% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentries_for_package.R
\name{get_bibentries_for_package}
\alias{get_bibentries_for_package}
\title{Get a citation file for an uninstalled Bioconductor package}
\usage{
get_bibentries_for_package(pkg, pkgversion, bioc_version = "devel")
}
\arguments{
\item{pkg}{The name of the package}

\item{pkgversion}{The version of Bioconductor to use}

\item{bioc_version}{The version of Bioconductor to use (default: devel)}
}
\value{
a bibentry object
}
\description{
This function downloads the tarball for a package and extracts the
CITATION file. It then parses the CITATION file and returns a bibentry
object, but hydrated with an extra field, pkg, that contains the name of
the package.
}
\examples{

library(BiocPkgTools)
pkgs = BiocPkgTools::biocPkgList()

biocgenerics_row = as.list(pkgs[pkgs$Package == "BiocGenerics",])

# should return NULL
res <- get_bibentries_for_package(biocgenerics_row$Package, biocgenerics_row$Version)
res

# should return a bibentry object
geoquery_row = as.list(pkgs[pkgs$Package == "GEOquery",])

res <- get_bibentries_for_package(geoquery_row$Package, geoquery_row$Version)
res
res$author
res$title
res$journal
res$doi
res$pkg

}
