% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_citations_for_package.R
\name{get_oa_works_for_bibentry}
\alias{get_oa_works_for_bibentry}
\title{get the openalex citation for a bibentry}
\usage{
get_oa_works_for_bibentry(bibentry)
}
\arguments{
\item{bibentry}{a bibentry object}
}
\value{
an openalex citation df
}
\description{
get the openalex citation for a bibentry
}
