% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_packages_through.R
\name{pkg_citation_list}
\alias{pkg_citation_list}
\title{Get package citation entries for all Bioconductor packages}
\usage{
pkg_citation_list(bioc_version = "devel")
}
\arguments{
\item{bioc_version}{The version of Bioconductor to use (default: devel)}
}
\value{
a \code{bibentry} object with each item named by package name.
Null entries are removed.
}
\description{
This function returns a list of bibentry objects, one for each Bioconductor
package. It uses the BiocPkgTools::biocPkgList function to get the list of
packages, and then calls get_bibentries_for_package for each package.
}
\examples{
\dontrun{
pkg_citations = pkg_citation_list()
pkg_citations$GEOquery
dois =
}
}
